requirements_airflow.txt
requests==2.31.0
pytest==8.2.1
google-cloud-storage==2.16.0
apache-airflow==2.9.1
pyspark==3.5.1
apache-airflow-providers-apache-spark
apache-airflow-providers-google==10.18.0
pandas
numpy
psycopg2-binary